requirements.txt
functions-framework==3.4.0
requests==2.31.0
psycopg2-binary==2.9.9
Flask==2.3.3
gunicorn==20.1.0
google-cloud-secret-manager